% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompARE-package.R
\docType{package}
\name{CompARE-package}
\alias{CompARE}
\alias{CompARE-package}
\title{CompARE: Statistical functions for studies with composite endpoints}
\description{
The package has been mainly designed for calculating the required sample size in randomized clinical trials with composite endpoints. This package also includes functions for calculating the probability of observing the composite endpoint and the expected effect on the composite endpoint.
}
\author{
\strong{Maintainer}: Marta Bofill Roig \email{marta.bofill.roig@upc.edu}

Authors:
\itemize{
  \item Jordi Cortes Martinez \email{jordi.cortes-martinez@upc.edu} (\href{https://orcid.org/0000-0002-3764-0795}{ORCID})
}

Other contributors:
\itemize{
  \item Guadalupe Gomez Melis (\href{https://orcid.org/0000-0003-4252-4884}{ORCID}) [contributor]
}

}
\keyword{internal}
