% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaSelection.R
\name{CopulaSelection}
\alias{CopulaSelection}
\title{Copula object}
\usage{
CopulaSelection(copula, rho, rhoType = "Spearman")
}
\arguments{
\item{copula}{Copula given from the following list: "Frank" (default), "Gumbel" , "Clayton", "Normal" , "T",
, "Galambos", "HuslerReiss", "Tawn", "Tev", "FGM" or "Plackett"}

\item{rho}{Spearman's coefficient between the 2 marginal distributions}

\item{rhoType}{type of correlation: "Spearman" (default) or "Kendall"}
}
\description{
Build a copula class object from the given family and the corresponding dependence parameter from the given correlation
}
\keyword{internal}
