% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sim.R
\name{f_sim}
\alias{f_sim}
\title{Simulation 1-arm}
\usage{
f_sim(samplesize, p_e1, p_e2, p_ce)
}
\arguments{
\item{samplesize}{sample size per arm}

\item{p_e1}{numeric parameter, probability of the event E1}

\item{p_e2}{numeric parameter, probability of the event E2}

\item{p_ce}{probability of the composite endpoint}
}
\value{
simulated data
}
\description{
Simulation of a trial with a composite endpoint
}
\examples{
data=f_sim(samplesize=100,p_e1=0.1,p_e2=0.5,p_ce=0.6)
head(data)
}
\keyword{internal}
