% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedman_formula.R
\name{freedman_formula}
\alias{freedman_formula}
\title{Freedman's formula}
\usage{
freedman_formula(alpha, power, HR)
}
\description{
Freedman's formula to calculate the number of events required
}
\keyword{internal}
