% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob1.R
\name{get_prob1}
\alias{get_prob1}
\title{Probability of observing the event}
\usage{
get_prob1(beta1, beta2, b11, b21, case, rho, copula = "Frank", endpoint)
}
\arguments{
\item{beta1}{Shape parameter for a Weibull law for the relevant event}

\item{beta2}{Shape parameter for a Weibull law for the additional event}

\item{b11}{Scale parameter of the Weibull distribution in treated arm for the relevant event}

\item{b21}{Scale parameter of the Weibull distribution in treated arm for the additional event}

\item{case}{Censoring case: 1, 2, 3 or 4}

\item{rho}{Spearman's coefficient between the 2 marginal distributions}

\item{copula}{Copula to use}

\item{endpoint}{Endpoint from which it is wanted to obtain the probability of being observed}
}
\description{
Returns the probability of observing a single endpoint
}
\keyword{internal}
