% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalsSelection.R
\name{MarginalsSelection}
\alias{MarginalsSelection}
\title{Marginal distributions for both components}
\usage{
MarginalsSelection(
  beta1,
  beta2,
  HR1,
  HR2,
  p1,
  p2,
  case,
  rho,
  theta,
  copula = "Frank",
  seed = 12345
)
}
\arguments{
\item{beta1}{Shape parameter for a Weibull law for the relevant event}

\item{beta2}{Shape parameter for a Weibull law for the additional event}

\item{HR1}{Hazard Ratio for a Weibull law for the relevant event}

\item{HR2}{Hazard Ratio for a Weibull law for the additional event}

\item{p1}{Proportion of the relevant event expected in group zero}

\item{p2}{Proportion of the additional event expected in group zero}

\item{case}{Censoring case}

\item{theta}{Dependence parameter for the bivariate distribution in control group}

\item{copula}{Copula to use}

\item{seed}{Seed for obtaining the probabilities of observing the events by simulation}
}
\description{
Returns the distribution and parameters of the marginal Weibull distributions
}
\keyword{internal}
