% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoendfeld_formula.R
\name{schoendfeld_formula}
\alias{schoendfeld_formula}
\title{Schoendfeld's formula}
\usage{
schoendfeld_formula(alpha, power, HR)
}
\description{
Schoendfeld's formula to calculate the number of events required
}
\keyword{internal}
