% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenfeld_formula.R
\name{schoenfeld_formula}
\alias{schoenfeld_formula}
\title{Schoenfeld's formula}
\usage{
schoenfeld_formula(alpha, power, HR)
}
\description{
Schoenfeld's formula to calculate the number of events required
}
\keyword{internal}
