% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SStar.R
\name{Sstar}
\alias{Sstar}
\title{Survival function for composite endpoint}
\usage{
Sstar(x, dist1, dist2, param1, param2, dist_biv)
}
\arguments{
\item{x}{Point in which to be evaluated}

\item{dist1}{Distribution function of the marginal T1 (pweibull)}

\item{dist2}{Distribution function of the marginal T2 (pweibull)}

\item{param1}{Parameters of the marginal distribution function T1 (pweibull)}

\item{param2}{Parameters of the marginal distribution function T2 (pweibull)}

\item{dist_biv}{Distribution function of the bivariate distribution via copula}
}
\description{
Returns the value of the survival function of S* at point x given the marginal distributions 
and the bivariate distributions via copula
}
\keyword{internal}
