% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sstar_func_perc.R
\name{Sstar_func_perc}
\alias{Sstar_func_perc}
\title{Difference between survival function and a given percentile}
\usage{
Sstar_func_perc(z, perc, ...)
}
\description{
Given a percentile perc, returns the difference between this percentile and the value of survival function
Useful for uniroot function
}
\keyword{internal}
