% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_corr.R
\name{upper_corr}
\alias{upper_corr}
\title{Upper bound for Pearson's Correlation}
\usage{
upper_corr(p_e1, p_e2)
}
\arguments{
\item{p_e1}{numeric parameter, probability of the event E1}

\item{p_e2}{numeric parameter, probability of the event E2}
}
\value{
Returns the maximum value that the correlation between the two outcomes can take.
}
\description{
Pearson's correlation between two binary outcomes takes values between two bounds defined according to the probabilities of the binary outcomes.
This function calculates the upper bound of the correlation based on the probabilities of two binary outcomes.
}
\details{
upper_corr returns a numeric value between 0 and 1.
}
\examples{
CompAREdesign::upper_corr(p_e1=0.3, p_e2=0.6)

}
