% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompARE-package.R
\docType{package}
\name{CompAREdesign-package}
\alias{CompAREdesign}
\alias{CompAREdesign-package}
\title{CompAREdesign: Statistical Functions for the Design of Studies with Composite Endpoints}
\description{
The package has been designed to calculate the required sample size in randomized clinical trials with composite endpoints. This package also includes functions to calculate the probability of observing the composite endpoint and the expected effect on the composite endpoint, among others.
}
\author{
\strong{Maintainer}: Marta Bofill Roig \email{marta.bofill.roig@upc.edu} (\href{https://orcid.org/0000-0002-4400-7541}{ORCID})

Authors:
\itemize{
  \item Jordi Cortes Martinez \email{jordi.cortes-martinez@upc.edu} (\href{https://orcid.org/0000-0002-3764-0795}{ORCID})
}

Other contributors:
\itemize{
  \item Guadalupe Gomez Melis (\href{https://orcid.org/0000-0003-4252-4884}{ORCID}) [contributor]
}

}
\keyword{internal}
